pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-agent-factory"
version = "0.1.0"
description = "Autonomous AI agent factory for crypto trading strategy R&D"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "ccxt>=4.4.0",
    "vectorbt>=0.26.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "statsmodels>=0.14.0",
    "typer[all]>=0.12.0",
    "rich>=13.9.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "httpx>=0.27.0",
    "orjson>=3.10.0",
    "tabulate>=0.9.0",
    "plotly>=5.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
agentfactory = "src.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
